Getting ready to predict golden json
