+++ petal_snorkel_predict_alex.py
@@ -1,5 +1,8 @@
 '''
-    Predicts what label something is based on 
+    Predicts what label something is based on trained data from golden json. 
+    This prediction doesn't really work at all. I think snorkel needs to be used on what it is trained on. 
+    If you have a new dataset then you should add it to the existing and run the predictions after training. 
+
     This example shows how to use the consensus label model approach to predict the "label" of any given text using snorkel
 
 '''
@@ -19,7 +22,7 @@ import pandas as pd
 from tqdm import trange
 import numpy as np 
 from ast import literal_eval
-from utils import load_dataset, evaluate_model,normalize_L
+from utils import load_dataset,normalize_L, single_model_to_dict, compare_single_model_dicts
 from create_labeling_functions import create_labeling_functions
 import ast
 '''
@@ -58,73 +61,6 @@ if osp.exists(large_model):
         global_translator_str = large_model_data['global_translator_str']
 
         large_model_L = normalize_L(L=L_match,translator=global_translator)
-
-def single_model_to_dict(label_model:LabelModel, translator_to_str:Dict[int,str],L_match:np.ndarray,model_index:int) -> List[Dict]:
-    """Evaluates the prediction accuracy of a single model
-
-    Args:
-        label_model (LabelModel): Model
-        translator_to_str (Dict[int,str]): dictionary containing keys -> string conversion
-        L_match (np.ndarray): this is the numpy array from pandas lf filter
-        model_index (int): index of model 
-
-    Returns:
-        List[Dict]: List of results
-    """
-    model = label_model
-    translator = translator_to_str
-    results_for_each_paper = list()
-    temp = evaluate_model(L_match, model, translator,model_index)
-    for p, paper in enumerate(temp):
-        probabilities = [t['probability'] for t in paper]
-        labels = [t['label'] for t in paper]
-        model_indicies = [t['model_index'] for t in paper]
-                
-        zipped = list(zip(probabilities,labels,model_indicies))
-        zipped = sorted(zipped, reverse=True)
-        probabilities, labels, model_indicies = zip(*zipped)
-    
-        results_for_each_paper.append({ 
-                'title':df.iloc[p]['title'], 'abstract':df.iloc[p]['abstract'],
-                'doi':df.iloc[p]['doi']})
-        if not pd.isna(df.iloc[p]['label_level_1']):
-            results_for_each_paper[p]['label'] = ast.literal_eval(df.iloc[p]['label_level_1'])
-        else:
-            results_for_each_paper[p]['label'] = 'not found'
-
-        results_for_each_paper[p]['label-snorkel-1'] = labels[0]
-        results_for_each_paper[p]['label-snorkel-2'] = labels[1]
-        results_for_each_paper[p]['label-snorkel-3'] = labels[2]
-        results_for_each_paper[p]['probability-snorkel-1'] = probabilities[0]
-        results_for_each_paper[p]['probability-snorkel-2'] = probabilities[1]
-        results_for_each_paper[p]['probability-snorkel-3'] = probabilities[2]
-        results_for_each_paper[p]['model-index-snorkel-1'] = model_indicies[0]
-        results_for_each_paper[p]['model-index-snorkel-2'] = model_indicies[1]
-        results_for_each_paper[p]['model-index-snorkel-3'] = model_indicies[2]
-    return results_for_each_paper
-
-def compare_single_model_dicts(result1:List[Dict], result2:List[Dict]) -> List[Dict]: 
-    """Compares one result with another and outputs the best case
-
-    Args:
-        result1 (List[Dict]): [description]
-        result2 (List[Dict]): [description]
-
-    Returns:
-        List[Dict]: [description]
-    """
-    for i in range(len(result1)):
-
-        if result2[i]['probability-snorkel-1'] > result1[i]['probability-snorkel-1']: 
-            result1[i]['label-snorkel-1']  = result2[i]['label-snorkel-1']
-            result1[i]['label-snorkel-1']  = result2[i]['label-snorkel-1']
-        if result2[i]['probability-snorkel-2'] > result1[i]['probability-snorkel-2']: 
-            result1[i]['label-snorkel-2']  = result2[i]['label-snorkel-2']
-            result1[i]['label-snorkel-2']  = result2[i]['label-snorkel-2']
-        if result2[i]['probability-snorkel-3'] > result1[i]['probability-snorkel-3']: 
-            result1[i]['label-snorkel-3']  = result2[i]['label-snorkel-3']
-            result1[i]['label-snorkel-3']  = result2[i]['label-snorkel-3']
-    return result1
 
 '''
     Evaluation using smaller models
